+++ lib/shopViewdPlugin.class.php
@@ -113,24 +113,15 @@ JS;
 
         try {
             if (($content_id === 'prices') && ($product instanceof shopProduct) && $product->id) {
-                $views = (int)$product->total_views;
-                if ($views) {
-                    $views = _wp("Просмотров карточки товара: ") . "<b>$views</b>";
-                } else {
-                    $views = _wp("Просмотров карточки товара пока не было");
-                }
-                $header = _wp("Количество просмотров карточки товара");
-                $content = <<<HTML
-<div class="s-content-section s-content-section__viewd">
-    <div class="s-section-header">
-        <h4 class="s-title">$header</h4>
-    </div>
-    <div class="s-section-body">
-        <p>$views</p>
-    </div>
-</div>
-HTML;
-                return ['bottom' => $content];
+                $view = wa()->getView();
+                $view->assign([
+                    'product_id' => (int)$product->id,
+                    'views'      => (int)$product->total_views,
+                    'header'     => _wp('Количество просмотров карточки товара'),
+                    'label'      => _wp('Просмотров карточки товара:'),
+                ]);
+                $template = $this->path . '/templates/hooks/backend_prod_content.html';
+                return ['bottom' => $view->fetch($template)];
             }
         } catch (Exception $e) {
             //do nothing